Header: Provide NOCATS option
Some FDOS programs might want to be built without kitten to reduce their
file size, obviously at the expense of translations being available.
Those can now be built as such by adding -DNOCATS to the CFLAGS, but it
is important to note that the kitten object should not be linked in as
not all (many?) linkers support smart linking.